refactor(results): decompose ResultDetails into smaller components
