+++ webapp/src/components/App/Singlestudy/explore/Results/ResultDetails/index.tsx
@@ -12,109 +12,37 @@
  * This file is part of the Antares project.
  */
 
-import DataGridViewer from "@/components/common/DataGridViewer";
 import { Column } from "@/components/common/Matrix/shared/constants";
-import { isNonEmptyMatrix, type ResultMatrixDTO } from "@/components/common/Matrix/shared/types";
-import ViewWrapper from "@/components/common/page/ViewWrapper";
+import { type ResultMatrixDTO } from "@/components/common/Matrix/shared/types";
 import useThemeColorScheme from "@/hooks/useThemeColorScheme";
-import GridOffIcon from "@mui/icons-material/GridOff";
-import { Box, Skeleton, Tab, Tabs } from "@mui/material";
-import React, { useCallback, useEffect, useMemo, useState, useRef } from "react";
-import { useTranslation } from "react-i18next";
+import type { Area, LinkElement, StudyMetadata } from "@/types/types";
+import React, { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import { useNavigate, useOutletContext, useParams } from "react-router";
 import usePromise from "../../../../../../hooks/usePromise";
-import useStudyOutput from "../hooks/useStudyOutput";
 import useAppSelector from "../../../../../../redux/hooks/useAppSelector";
 import { getAreas, getLinks } from "../../../../../../redux/selectors";
 import { getStudyMatrixIndex } from "../../../../../../services/api/matrix";
 import { getStudyData } from "../../../../../../services/api/study";
-import type { Area, LinkElement, StudyMetadata } from "../../../../../../types/types";
-import { toError } from "../../../../../../utils/fnUtils";
 import { isSearchMatching } from "../../../../../../utils/stringUtils";
-import ButtonBack from "../../../../../common/ButtonBack";
-import FilterableMatrixGrid, {
-  type FilterableMatrixGridHandle,
-} from "../../../../../common/Matrix/components/FilterableMatrixGrid";
 import {
-  generateCustomColumns,
   generateDateTime,
   generateResultColumns,
   groupResultColumns,
 } from "../../../../../common/Matrix/shared/utils";
-import EmptyView from "../../../../../common/page/EmptyView";
-import PropertiesView from "../../../../../common/PropertiesView";
 import SplitView from "../../../../../common/SplitView/index";
-import UsePromiseCond from "../../../../../common/utils/UsePromiseCond";
-import ListElement from "../../common/ListElement";
-import ResultFilters from "./ResultFilters";
-import { createPath, DataType, MAX_YEAR, OutputItemType, SYNTHESIS_ITEMS, Timestep } from "./utils";
+import useStudyOutput from "../hooks/useStudyOutput";
+import { DataType, OutputItemType, SYNTHESIS_ITEMS, Timestep, createPath } from "./utils";
+import ResultItemSelector from "./components/ResultItemSelector";
+import ResultMatrixViewer from "./components/ResultMatrixViewer";
+import SynthesisViewer, { type SynthesisData } from "./components/SynthesisViewer";
+import type { FilterableMatrixGridHandle } from "@/components/common/Matrix/components/FilterableMatrixGrid";
 
 type SetResultColHeaders = (headers: string[][], indices: number[]) => void;
-
-interface LeftPanelProps {
-  itemType: OutputItemType;
-  onItemTypeChange: (_event: React.SyntheticEvent, newValue: OutputItemType) => void;
-  output: { synthesis?: boolean } | undefined;
-  filteredItems: Array<{ id: string; name: string; label?: string }>;
-  selectedItemId: string;
-  onSetSelectedItemId: (item: { id: string }) => void;
-  onSearchChange: (value: string) => void;
-  onNavigateBack: () => void;
-}
-
-const LeftPanel = React.memo(function LeftPanel({
-  itemType,
-  onItemTypeChange,
-  output,
-  filteredItems,
-  selectedItemId,
-  onSetSelectedItemId,
-  onSearchChange,
-  onNavigateBack,
-}: LeftPanelProps) {
-  const { t } = useTranslation();
-
-  return (
-    <Box>
-      <PropertiesView
-        topContent={
-          <Box sx={{ width: 1, px: 1 }}>
-            <ButtonBack onClick={onNavigateBack} />
-          </Box>
-        }
-        mainContent={
-          <>
-            <Tabs
-              value={itemType}
-              onChange={onItemTypeChange}
-              size="extra-small"
-              variant="fullWidth"
-            >
-              <Tab label={t("study.areas")} value={OutputItemType.Areas} />
-              <Tab label={t("study.links")} value={OutputItemType.Links} />
-              {output?.synthesis && (
-                <Tab label={t("study.synthesis")} value={OutputItemType.Synthesis} />
-              )}
-            </Tabs>
-            <ListElement
-              list={filteredItems}
-              currentElement={selectedItemId}
-              currentElementKeyToTest="id"
-              setSelectedItem={onSetSelectedItemId}
-            />
-          </>
-        }
-        onSearchFilterChange={onSearchChange}
-      />
-    </Box>
-  );
-});
 
 function ResultDetails() {
   const { study } = useOutletContext<{ study: StudyMetadata }>();
   const { isDarkMode } = useThemeColorScheme();
   const { outputId } = useParams();
-  const { t } = useTranslation();
   const navigate = useNavigate();
 
   const [dataType, setDataType] = useState(DataType.General);
@@ -123,6 +51,9 @@ function ResultDetails() {
   const [itemType, setItemType] = useState(OutputItemType.Areas);
   const [selectedItemId, setSelectedItemId] = useState("");
   const [searchValue, setSearchValue] = useState("");
+  // Store filtered headers and their original indices separately
+  // This allows us to correctly map the data rows to their corresponding headers
+  // when some columns are filtered out
   const [resultColHeaders, setResultColHeaders] = useState<string[][]>([]);
   const [headerIndices, setHeaderIndices] = useState<number[]>([]);
 
@@ -144,7 +75,6 @@ function ResultDetails() {
       label?: string;
     }>;
 
-    // Only return if items actually exist and are valid
     return Array.isArray(currentItems) ? currentItems : [];
   }, [itemType, areas, links]);
 
@@ -237,7 +167,7 @@ function ResultDetails() {
     return matrixRes.data.data.map((row) => headerIndices.map((index) => row[index]));
   }, [matrixRes.data, headerIndices]);
 
-  const synthesisRes = usePromise(
+  const synthesisRes = usePromise<SynthesisData | null>(
     () => {
       if (outputId && selectedItem && isSynthesis) {
         const path = `output/${outputId}/economy/mc-all/grid/${selectedItem.id}`;
@@ -324,7 +254,7 @@ function ResultDetails() {
 
   return (
     <SplitView id="results" sizes={[15, 85]} minSize={[200, 300]}>
-      <LeftPanel
+      <ResultItemSelector
         itemType={itemType}
         onItemTypeChange={handleItemTypeChange}
         output={output}
@@ -335,90 +265,30 @@ function ResultDetails() {
         onNavigateBack={handleNavigateBack}
       />
 
-      <ViewWrapper flex>
-        {isSynthesis ? (
-          <UsePromiseCond
-            response={synthesisRes}
-            ifPending={() => <Skeleton sx={{ height: 1, transform: "none" }} />}
-            ifFulfilled={(matrix) => {
-              if (!matrix) {
-                return <EmptyView title={t("study.results.noData")} icon={GridOffIcon} />;
-              }
-
-              return (
-                <DataGridViewer
-                  data={matrix.data}
-                  columns={generateCustomColumns({
-                    titles: matrix.columns,
-                  })}
-                />
-              );
-            }}
-          />
-        ) : (
-          <Box sx={{ display: "flex", flexDirection: "column", height: 1, width: 1 }}>
-            <Box sx={{ flexShrink: 0 }}>
-              <ResultFilters
-                year={year}
-                setYear={setYear}
-                dataType={dataType}
-                setDataType={setDataType}
-                timestep={timestep}
-                setTimestep={setTimestep}
-                maxYear={output?.nbyears || MAX_YEAR}
-                studyId={study.id}
-                path={path}
-                colHeaders={matrixRes.data?.columns ?? resultColHeaders}
-                onColHeadersChange={handleColHeadersChange}
-                onToggleFilter={handleToggleFilter}
-              />
-            </Box>
-            <Box sx={{ flex: 1, minHeight: 0, display: "flex", flexDirection: "column" }}>
-              <UsePromiseCond
-                response={matrixRes}
-                ifPending={() => <Skeleton sx={{ height: 1, transform: "none" }} />}
-                ifFulfilled={() => {
-                  // If headers haven't been processed yet, show loading
-                  if (resultColHeaders.length === 0) {
-                    return <Skeleton sx={{ height: 1, transform: "none" }} />;
-                  }
-
-                  // If we have processed headers but no valid data, show "No data"
-                  if (!isNonEmptyMatrix(filteredData)) {
-                    return <EmptyView title={t("study.results.noData")} icon={GridOffIcon} />;
-                  }
-
-                  return (
-                    <FilterableMatrixGrid
-                      ref={matrixGridRef}
-                      key={`grid-${resultColHeaders.length}`}
-                      data={filteredData}
-                      rows={filteredData.length}
-                      columns={resultColumns}
-                      dateTime={dateTime}
-                      timeFrequency={dateTimeMetadata?.level}
-                      readOnly
-                    />
-                  );
-                }}
-                ifRejected={(err) => (
-                  <EmptyView
-                    title={
-                      // 404 error is expected when their is no data
-                      // for the selected area or link result
-                      // TODO: Instead this should be an empty response from the server
-                      toError(err).message.includes("404")
-                        ? t("study.results.noData")
-                        : t("data.error.matrix")
-                    }
-                    icon={GridOffIcon}
-                  />
-                )}
-              />
-            </Box>
-          </Box>
-        )}
-      </ViewWrapper>
+      {isSynthesis ? (
+        <SynthesisViewer synthesisRes={synthesisRes} />
+      ) : (
+        <ResultMatrixViewer
+          matrixRes={matrixRes}
+          resultColHeaders={resultColHeaders}
+          filteredData={filteredData}
+          resultColumns={resultColumns}
+          matrixGridRef={matrixGridRef}
+          dateTime={dateTime}
+          dateTimeMetadata={dateTimeMetadata}
+          year={year}
+          setYear={setYear}
+          dataType={dataType}
+          setDataType={setDataType}
+          timestep={timestep}
+          setTimestep={setTimestep}
+          output={output}
+          studyId={study.id}
+          path={path}
+          onColHeadersChange={handleColHeadersChange}
+          onToggleFilter={handleToggleFilter}
+        />
+      )}
     </SplitView>
   );
 }
